requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
Flask
Flask-Login
Flask-SQLAlchemy
gunicorn
python-dotenv
Werkzeug
mercadopago
flask-socketio
python-engineio
python-socketio
